+++ CHANGES.md
@@ -1,4 +1,6 @@
 [unreleased]
+* remove constructor
+* add some tests
 
 #### 0.5.0 / 2021-03-12
 * de-anonymize hooks